+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
   * Protocol: Delgado, Delgado-Reusable, Delgado-Library (#1)
   * Protocol: Support for SmartJudge (#20)
-  * Renderer: Removed renderers `yaml` and `json`
+  * Renderer: Removed renderers `yaml`, `json` and `human-readable`
   * Renderer: Created new renderers `game-matrix` and `game-tree` (covers #33)
   * Feature: Added event filtering support for environments
   * Feature: Added `--protocol-path` parameter for limiting protocol paths to be executed (#23)